Move 'OpenLayers Map' component to appropriate package
