--- client/src/style/openlayers-styles.css
@@ -1,8 +1,0 @@
-.openlayers-map {
-  height: 100%;
-}
-
-.map-container {
-  max-height: 100%;
-  max-width: 100%;
-}
